Modifica rotas de petiano em tutor, corrige a issue #63 e parte do #44
