+++ pages/tutor/eventos/organizadores/create.vue
@@ -6,20 +6,37 @@
     </b-alert>
     <div class="card">
       <div class="card-header">
-        <strong><i class="fa fa-edit"></i> Organizadores</strong> <small>Formulário de criação</small>
+        <strong>
+          <i class="fa fa-edit"></i> Organizadores
+        </strong>
+        <small>Formulário de criação</small>
         <div class="card-actions">
-          <a href="/petiano/rh/organizadores/" class="btn btn-close"><i class="icon-close"></i></a>
+          <nuxt-link to="/tutor/eventos/organizadores" class="btn btn-close">
+            <i class="icon-close"></i>
+          </nuxt-link>
         </div>
       </div>
       <div class="card-body">
         <form @submit="submitForm">
           <div class="form-group">
-            <label for="exampleFormControlInput1">ID Pessoa:</label>
-            <input type="number" class="form-control" v-model="form.pessoa" />
+            <label for="exampleFormControlInput1">Pessoa:</label>
+            <select class="form-control" v-model="form.pessoa">
+              <option
+                v-for="participante in pessoas"
+                :key="participante.idPessoa"
+                :value="participante.idPessoa"
+              >{{ participante.nome }}</option>
+            </select>
           </div>
           <div class="form-group">
-            <label for="exampleFormControlInput1">ID Evento:</label>
-            <input type="number" class="form-control" v-model="form.evento" />
+            <label for="exampleFormControlInput1">Evento:</label>
+            <select class="form-control" v-model="form.evento">
+              <option
+                v-for="evento in eventos"
+                :key="evento.idEvento"
+                :value="evento.idEvento"
+              >{{ evento.titulo }}</option>
+            </select>
           </div>
           <!--<div class="form-group">
             <label for="exampleFormControlInput1">confirmado</label>
@@ -30,8 +47,12 @@
             <input type="bool" class="form-control" v-model="form.espera" />
           </div>-->
           <div class="form-group">
-            <b-button type="submit" variant="primary"><i class="fa fa-dot-circle-o"></i> Enviar</b-button>
-            <b-button type="reset" variant="danger"><i class="fa fa-ban"></i> Limpar campos</b-button>
+            <b-button type="submit" variant="primary">
+              <i class="fa fa-dot-circle-o"></i> Enviar
+            </b-button>
+            <b-button type="reset" variant="danger">
+              <i class="fa fa-ban"></i> Limpar campos
+            </b-button>
           </div>
         </form>
       </div>
@@ -42,7 +63,7 @@
 import axios from "~/axios";
 
 export default {
-  layout: "menu/petiano",
+  layout: "menu/tutor",
   data() {
     return {
       form: {
@@ -52,34 +73,47 @@ export default {
       alert: {
         message: "",
         class: ""
-      }
+      },
+      pessoas: {},
+      eventos: {}
     };
   },
   methods: {
     submitForm(e) {
       axios
-        .post("organizadores-cadastrar/" + this.form.evento + "/" + this.form.pessoa)
+        .post(
+          "organizadores-cadastrar/" + this.form.evento + "/" + this.form.pessoa
+        )
         .then(res => {
           this.alert.class = "success";
           this.alert.message = "Organizador cadastrado com sucesso.";
           this.form = Object.entries(this.form).map(item => {
             return (item = "");
-          this.$router.push({ path : '/petiano/rh/organizadores/' });
-
+            this.$router.push({ path: "/tutor/eventos/organizadores" });
           });
         })
         .catch(err => {
           this.alert.class = "danger";
-          this.alert.message = "Organizador não cadastrado. Por favor, tente novamente";
+          this.alert.message =
+            "Organizador não cadastrado. Por favor, tente novamente";
         });
       e.preventDefault();
     },
     onReset(evt) {
-      evt.preventDefault()
+      evt.preventDefault();
       // Reset our form values
-      this.form.pessoa = 0
-      this.form.evento = 0
+      this.form.pessoa = 0;
+      this.form.evento = 0;
     }
+  },
+  mounted() {
+    axios.get("pessoas").then(res => {
+      this.pessoas = res.data.content;
+    });
+
+    axios.get("eventos-abertos").then(res => {
+      this.eventos = res.data;
+    });
   }
 };
 </script>
