+++ pages/tutor/eventos/participantes/index.vue
@@ -79,10 +79,7 @@
 import axios from "~/axios";
 export default {
   name: "dashboard",
-  /* TODO:: Esse layout será apresentado tanto pro petiano quando pro coordenador
-  depois será necessário uma lógica pra chamar o layout dependendo do tipo de usuário
-  logado. No momento trabalharei apenas com os petianos. */
-  layout: "menu/petiano",
+  layout: "menu/tutor",
   data() {
     return {
       keyword: '',
