+++ pages/tutor/pessoas-cadastradas/index.vue
@@ -4,28 +4,37 @@
       <template v-slot:header>
         <h3>Pessoas cadastradas</h3>
       </template>
+      <div v-if="pessoas.length > 0">
+        <b-input-group class="mt-3 mb-3">
+          <b-form-input placeholder="Busca" type="text" v-model="keyword"></b-form-input>
+          <b-input-group-append>
+            <b-button variant="success" @click="search">
+              <i class="fa fa-search"></i>
+            </b-button>
+            <b-button variant="outline-danger" @click="getPessoas">
+              <i class="fa fa-remove"></i>
+            </b-button>
+          </b-input-group-append>
+        </b-input-group>
 
-      <b-input-group  class="mt-1 mb-3" >
-        <!-- Always bind the id to the input so that it can be focused when needed -->
-        <b-form-input
-          v-model="keyword"
-          placeholder="Busca por nome ou por CPF"            
-          type="text"
-        ></b-form-input>
-        <b-input-group-text slot="append">
-          <b-btn class="p-0" :disabled="!keyword" variant="link" size="sm" @click="keyword = ''"><i class="fa fa-remove"></i></b-btn>
-      </b-input-group-text>
-      </b-input-group>
-
-      <div  v-if="pessoas.length > 0">
         <b-table
           responsive="sm"
-          :items="items"
+          :items="pessoas"
           :current-page="currentPage"
           :bordered="true"
-          :per-page="10"
+          :per-page="20"
           :fields="fields"
         >
+          <template v-slot:cell(tipo_usuario)="row">
+            <select @change="changePermission(row, $event)" class="form-control">
+              <option
+                v-for="tipo in tipos_usuario"
+                :key="tipo.idTipo_usuario"
+                :value="tipo.idTipo_usuario"
+                :selected="row.item.tipo_usuario.idTipo_usuario === tipo.idTipo_usuario"
+              >{{ tipo.nome }}</option>
+            </select>
+          </template>
           <template v-slot:cell(actions)="row">
             <b-button 
                 @click.prevent="editar(row.item.idPessoa, row.item.nome, row.item.cpf, row.item.tipo_usuario)" 
@@ -35,8 +44,8 @@
         </b-table>
         <nav>
           <b-pagination
-            :total-rows="items.length"
-            :per-page="10"
+            :total-rows="numPages"
+            :per-page="20"
             v-model="currentPage"
             prev-text="Anterior"
             next-text="Próximo"
@@ -44,55 +53,158 @@
           />
         </nav>
       </div>
-      <div v-else>Nenhum pessoa cadastrada</div>
+      <div v-else>Nenhuma pessoa cadastrada</div>
     </b-card>
+
+    <b-modal ref="modal-create" title="Informações adicionais" hide-footer no-close-on-backdrop>
+      <label for="data-ingresso">Data ingresso</label>
+      <b-form-datepicker
+        id="data-ingresso"
+        v-model="modal.data_ingresso"
+        type="date"
+        required
+        locale="pt-br"
+        label-no-date-selected="Nenhuma data selecionada"
+      ></b-form-datepicker>
+      <b-button variant="primary" @click="createOrUpdatePetiano('create')" class="float-right">OK</b-button>
+    </b-modal>
+
+    <b-modal ref="modal-update" title="Informações adicionais" hide-footer no-close-on-backdrop>
+      <label for="data-egresso">Data egresso</label>
+      <b-form-datepicker
+        id="data-egresso"
+        v-model="modal.data_egresso"
+        type="date"
+        required
+        locale="pt-br"
+        label-no-date-selected="Nenhuma data selecionada"
+      ></b-form-datepicker>
+      <b-button variant="primary" @click="createOrUpdatePetiano('update')" class="float-right">OK</b-button>
+    </b-modal>
   </div>
 </template>
 
 <script>
 import axios from "~/axios";
+import Cookies from "js-cookie";
 
 export default {
   name: "dashboard",
   layout: "menu/tutor",
   data() {
     return {
-      keyword: '',
-      pessoas: [],
+      pessoas: {},
+      tipos_usuario: {},
       currentPage: 1,
       fields: [
         { key: "nome", sortable: true },
-        { key: "cpf", sortable: true,label: "CPF", formatter: (value) => { if (value != null) return `${value.substring(0, 3)}.${value.substring(3, 6)}.${value.substring(6, 9)}-${value.substring(9, 11)}`}  },
-        { key: "actions", sortable: true, label: "Ações disponíveis" }
-      ]
+        { key: "cpf", sortable: false },
+        {
+          key: "tipo_usuario",
+          sortable: true,
+          label: "Tipo de usuário",
+          formatter: value => {
+            return value.nome;
+          }
+        },
+        {
+          key: "usuario",
+          sortable: true,
+          label: "Email",
+          formatter: value => {
+            return value.email;
+          }
+        },
+        { key: "actions", sortable: true, label: "Ações disponíveis" },
+      ],
+      modal: {},
+      keyword: "",
+      numPages: 1,
     };
   },
-  computed: {
-    items () {
-      return this.keyword
-          ? this.pessoas.filter(item => item.cpf.includes(this.keyword) || item.nome.includes(this.keyword))
-          : this.pessoas
-    }
-  },
   mounted() {
-    axios.get("pessoas").then(res => {
-      this.pessoas = res.data.content;
+    this.getPessoas();
+
+    axios.get("tipo-usuario").then(res => {
+      this.tipos_usuario = res.data;
     });
   },
+  watch: {
+    currentPage: function(val){
+      axios.get("pessoas?page=" + val).then(res => {
+        this.pessoas = res.data.content;
+        this.pageable = res.data.totalPages;
+      });
+    }
+  },
   methods: {
-    editar(idPessoa, nome, cpf, tipo_usuario){
-      this.$router.push({
-          path: 'edit/',
-          query  : {"idPessoa": idPessoa,
-                    "nome": nome,
-                    "cpf": cpf,
-                    "tipo_usuario":tipo_usuario}
+    async changePermission(row, event) {
+      let id_usuario = row.item.usuario.idUsuario;
+      let id_tipo = event.target.value;
+      await axios
+        .post(
+          "pessoas-cadastro-atualizar/" + id_tipo + "/" + id_usuario,
+          this.pessoas.filter(item => item.idPessoa === row.item.idPessoa)[0]
+        )
+        .then(res => {
+          if (id_tipo == 2) {
+            // se tipo petiano
+            this.$set(this.modal, "item", row.item);
+            this.showModal("modal-create");
+          } else if (id_tipo != 2) {
+            this.$set(this.modal, "item", row.item);
+            this.showModal("modal-update");
+          }
         })
+        .catch(err => {
+          console.log(err);
+        });
+    },
+    showModal(name) {
+      console.log("abre modal");
+      this.$refs[name].show();
+    },
+    hideModal(name) {
+      this.$refs[name].hide();
+    },
+    createOrUpdatePetiano(type) {
+      let pessoaSelected = this.pessoas.filter(
+        item => item.idPessoa === this.modal.item.idPessoa
+      )[0];
+      if (type === "create") {
+        pessoaSelected["data_ingresso"] = this.modal.data_ingresso;
+        pessoaSelected["data_egresso"] = null;
+      } else {
+        pessoaSelected["data_egresso"] = this.modal.data_egresso;
+      }
+
+      axios
+        .post("petianos-cadastro/" + this.modal.item.idPessoa, pessoaSelected)
+        .then(res => {
+          this.hideModal("modal-create");
+          this.hideModal("modal-update");
+        })
+        .catch(err => {
+          alert(
+            "Algo deu errado na hora de cadastrar o petiano. Tente novamente mais tarde!"
+          );
+        });
+    },
+    search() {
+      axios.get("pesquisar-pessoa/" + this.keyword).then(res => {
+        this.pessoas = res.data.content;
+      });
+    },
+    getPessoas(){
+      axios.get("pessoas").then(res => {
+        this.pessoas = res.data.content;
+        this.pageable = res.data.totalPages;
+
+      });
     }
   }
 };
 </script>
-
 
 <style scoped>
 h3 {
